// web/src/api/generated/models/checkin-ranking-item.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface CheckinRankingItem {
    'userId'?: number;
    'userName'?: string;
    'userAvatar'?: string;
    'totalCheckinDays'?: number;
    'currentStreak'?: number;
    'rank'?: number;
}